Added types for nuxt module
