+++ packages/core/nuxt-module/src/helpers/isProduction.ts
@@ -1,1 +1,3 @@
-export default (options) => process.env.NODE_ENV === 'production' || options.coreDevelopment;
+import { ModuleOptions } from '../types';
+
+export default (options: ModuleOptions): boolean => process.env.NODE_ENV === 'production' || options.coreDevelopment;
