+++ packages/core/nuxt-module/src/helpers/log.ts
@@ -1,7 +1,8 @@
 import consola from 'consola';
 import chalk from 'chalk';
+import { Log } from '../types';
 
-const log = Object.freeze({
+const log: Log = Object.freeze({
   info: (message) => consola.info(chalk.bold('VSF'), message),
   success: (message) => consola.success(chalk.bold('VSF'), message),
   warning: (message) => consola.warn(chalk.bold('VSF'), message),
