+++ packages/core/nuxt-module/src/helpers/resolveDependency.ts
@@ -1,4 +1,4 @@
-export default function resolveDependency(name) {
+export default function resolveDependency(name: string): string {
   try {
     return require.resolve(name, { paths: [process.cwd()] });
   } catch (error) {
