+++ packages/core/nuxt-module/src/modules/raw-sources-loader.ts
@@ -1,8 +1,9 @@
 import log from '../helpers/log';
 import isProduction from '../helpers/isProduction';
 import resolveDependency from '../helpers/resolveDependency';
+import { ModuleOptions } from '../types';
 
-export default function VueStorefrontPerformanceModule (options) {
+export default function VueStorefrontPerformanceModule(options: ModuleOptions): void {
   const useRawSource = (pckg) => {
     const pkgPath = resolveDependency(`${pckg}/package.json`);
     // eslint-disable-next-line global-require
