+++ packages/core/nuxt-module/src/modules/storefront-ui.ts
@@ -1,7 +1,8 @@
 import merge from '../helpers/merge';
+import { ModuleOptions } from '../types';
 
 // TODO: Create a separate nuxt module for storefront ui
-function loadStorefrontRawSources (options) {
+function loadStorefrontRawSources(options: ModuleOptions): void {
   const rawSources = [
     '@storefront-ui/vue',
     '@storefront-ui/shared'
@@ -13,6 +14,6 @@ function loadStorefrontRawSources (options) {
   });
 }
 
-export default function VueStorefrontPerformanceModule (options) {
+export default function VueStorefrontPerformanceModule(options: ModuleOptions): void {
   loadStorefrontRawSources.call(this, options);
 }
